+++ test/assignments/wizard_test.dart
@@ -1,5 +1,5 @@
-import 'package:modu_3_dart_study/wand.dart';
-import 'package:modu_3_dart_study/wizard.dart';
+import 'package:modu_3_dart_study/assignments/wand.dart';
+import 'package:modu_3_dart_study/assignments/wizard.dart';
 import 'package:test/test.dart';
 
 void main() {
